detector/requirements.txt
paho-mqtt==1.6.1
opencv-python-headless==4.8.1.78
numpy==1.24.3
